Fix bugs in day 6b
There were some mistakes in my solution. Essentially, the strat
of placing an obstacle at each timestep has the bug that it is
possible to pass through a square in one direction, and then later
attempt to put an obstacle there when you pass by it in another
direction. This is illegal because it would have affected the
direction you go in an earlier timestep.

To solve this, simply pre-compute the set of coords you need to
test based on the part-a traversal of the path.